% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\docType{data}
\name{acs_codes}
\alias{acs_codes}
\alias{acs_codes_long}
\alias{acs_sex_codes}
\alias{acs_bpl_codes}
\alias{acs_educ_codes}
\title{Encodings for 2018 ACS microdata}
\format{A data frame.}
\source{
\href{https://usa.ipums.org/usa/index.shtml}{IPUMS-USA}
}
\usage{
acs_codes

acs_codes_long

acs_sex_codes

acs_bpl_codes

acs_educ_codes
}
\description{
Encodings for example dataset based on the American Community Survey in 2018.
}
\seealso{
\link{acs_nh}

Package \link[ipumsr:ipumsr-package]{ipumsr} to get real census microdata
}
\keyword{datasets}
