% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\docType{data}
\name{acs_nh}
\alias{acs_nh}
\title{Synthetic 2018 ACS microdata for New Hampshire}
\format{A data frame.
\describe{
\item{PERWT}{weight}
\item{SEX}{encoded sex/gender}
\item{AGE}{age in years}
\item{RACE}{encoded race}
\item{HISPAN}{encoded Hispanic ethnicity}
\item{BPL}{encoded birth place: country or US state or territory}
\item{EDUC}{encoded level of education}
}}
\source{
Synthesized using data from
\href{https://usa.ipums.org/usa/index.shtml}{IPUMS-USA}
}
\usage{
acs_nh
}
\description{
Example dataset based on the American Community Survey of the state of
New Hampshire in 2018. \emph{\strong{This is not real data.}} It was synthesized using
\code{\link[synthpop:syn]{synthpop::syn()}}. This data should only be used for examples and cannot be
used in real analyses.
}
\seealso{
\link{acs_codes}

Package \link[ipumsr:ipumsr-package]{ipumsr} to get real census microdata

Package \link[synthpop:synthpop-package]{synthpop} to synthesize data
}
\keyword{datasets}
