% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_list_of_lists.R
\name{is_list_of_lists}
\alias{is_list_of_lists}
\title{Test if an input is a list of lists}
\usage{
is_list_of_lists(list)
}
\arguments{
\item{...}{Character vectors of two or more columns names to interact,
optionally in a list}
}
\value{
A logical value
}
\description{
Description
}
\seealso{
\code{\link[=is.list]{is.list()}} to test if an input is a list\preformatted{[purrr::flatten()] to remove a level of hierarchy from a list

[purrr::squash()] to remove all levels of hierarchy from a list
}
}
