% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_interact.R
\name{leaf_interact_all}
\alias{leaf_interact_all}
\title{Create an interaction column}
\usage{
leaf_interact_all(
  tbl,
  ...,
  cols = NULL,
  except = NULL,
  sep = " x ",
  col_sep = NULL
)
}
\arguments{
\item{tbl}{A data frame}

\item{...}{One or more column names to interact with all other columns,
either unquoted names or a character vector}

\item{cols}{A string or character vector of one or more column names}

\item{except}{An unquote name or a character vector of one or more column
names not to interact, e.g. a column of weights}

\item{sep}{A string indicating the separator to place between values in
observations}

\item{col_sep}{A string indicating the separator to place between column
names. If NULL, defaults to \code{sep} with spaces replaced by underscores}
}
\value{
A modified \code{tbl} with an interaction column added
}
\description{
Description
}
\seealso{
\code{\link[=leaf_interact]{leaf_interact()}} to create interactions one at a time\preformatted{[leaf_interactions()] to create multiple interactions at once
}
}
