% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_interact.R
\name{leaf_interactions}
\alias{leaf_interactions}
\title{Create interaction columns}
\usage{
leaf_interactions(tbl, ..., sep = " x ", col_sep = NULL)
}
\arguments{
\item{tbl}{A data frame}

\item{...}{Character vectors of two or more columns names to interact,
optionally in a list}

\item{sep}{A string indicating the separator to place between values in
observations}

\item{col_sep}{A string indicating the separator to place between column
names. If NULL, defaults to \code{sep} with spaces replaced by underscores}
}
\value{
A modified \code{tbl} with an interaction column added
}
\description{
Description
}
\seealso{
\code{\link[=leaf_interact]{leaf_interact()}} to create interactions one at a time\preformatted{[leaf_interact_all()] to create interactions bewteen a column and all
other columns
}
}
