% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_other.R
\name{leaf_other}
\alias{leaf_other}
\title{Consolidate levels into other category}
\usage{
leaf_other(tbl, cutoff, other = "Other", inclusive = TRUE)
}
\arguments{
\item{tbl}{A data frame with three columns -- \code{variable}, \code{level}, and
\code{proportion}, as created by \code{leaf_peep()}}

\item{cutoff}{The proportion below which levels will be turned to other}

\item{other}{A string indicating the name to be used for the other category,
defaults to "Other"}

\item{inclusive}{If TRUE, when a variable's other category is still below the
cutoff, the next smallest level will also be converted to other.
If FALSE, the other category may remain below the cutoff}
}
\value{
A data frame where levels with proportions below the cutoff are
consolidated into an "other" category
}
\description{
Consolidate levels into other category
}
\examples{
# Set up an example data frame with some very small levels
df <- dplyr::tibble(
  variable   = c(rep("letter", 4), rep("number", 4)),
  level      = c(letters[1:4], 1:4),
  proportion = c(0.6, 0.395, 0.0045, 0.0045, 0.6, 0.395, 0.0055, 0.0035)
)

# Consolidate small categories so that all values are above the cutoff
leaf_other(df, cutoff = 0.005)

# You can give a different name to the other category
leaf_other(df, cutoff = 0.005, other = "REDACTED")
leaf_other(df, cutoff = 0.005, other = NA)

# If the other category is smaller than the cutoff, leaf_other() will convert
# the next smallest value to other, even if it is larger than the cutoff.
# Use `inclusive = FALSE` to only convert categories smaller than the cutoff.
leaf_other(df, cutoff = 0.005, inclusive = FALSE)
}
