% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_peep.R
\name{leaf_peep}
\alias{leaf_peep}
\title{Convert a data frame to weighting targets}
\usage{
leaf_peep(tbl, weight_col = NULL)
}
\arguments{
\item{tbl}{A data frame of individual observations, e.g. census microdata}

\item{weight_col}{A column of individual weights, as an unquoted name or
character string}
}
\value{
A data frame of weighting targets
}
\description{
Convert a data frame to weighting targets
}
\examples{
# Load a dataset of example microdata
acs_md

# Generate weighting targets
leaf_peep(acs_md)

# Microdata with a weights column
leaf_peep(acs_md, weight_col = "PERWT")
}
