% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_join.R
\name{recode_join}
\alias{recode_join}
\title{Recode a column of a data frame by joining with a data frame of values}
\usage{
recode_join(tbl, col, code_tbl, code_col = code)
}
\arguments{
\item{tbl}{Data frame to modify}

\item{col}{The column to recode}

\item{code_tbl}{A data frame containing a column of codes and a column of
replacment values with the same name as \code{col}}

\item{code_col}{The column of \code{code_tbl} containing the codes present in
\code{tbl}}
}
\value{
A data frame with a selected column recoded
}
\description{
Recode a column of a data frame by joining with a data frame of values
}
\examples{
raw_data <- tibble(race = runif(1000, 1, 4), answer = runif(1000, 1, 4)) \%>\%
  mutate_all(round)

race_codes <- tribble(
  ~ code, ~ race,
  1, "AAPI",
  2, "Black",
  3, "Latinx",
  4, "White"
)

answer_codes <- tribble(
  ~ code, ~ answer,
  1, "Agree",
  2, "Disagree",
  3, "No opinion",
  4, "Not sure"
)

raw_data \%>\%
  recode_join(race, race_codes, code) \%>\%
  recode_join(answer, answer_codes, code)
}
