% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_md.R
\docType{data}
\name{acs_codes}
\alias{acs_codes}
\alias{acs_codes_long}
\alias{acs_race_codes}
\alias{acs_age_codes}
\title{Encodings for 2018 ACS microdata}
\format{A data frame.}
\source{
\href{https://usa.ipums.org/usa/index.shtml}{IPUMS-USA}
}
\usage{
acs_codes

acs_codes_long

acs_race_codes

acs_age_codes
}
\description{
Encodings for example dataset based on the American Community Survey in 2018.
}
\seealso{
Package \link[ipumsr:ipumsr-package]{ipumsr} to get real census
microdata
}
\keyword{datasets}
