% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_interact.R
\name{leaf_interact}
\alias{leaf_interact}
\title{Create an interaction column}
\usage{
leaf_interact(tbl, ..., cols, sep = " x ", col_sep = NULL)
}
\arguments{
\item{tbl}{A data frame}

\item{...}{Two or more column names to interact, optionally in a character
vector}

\item{cols}{A character vector of two or more column names}

\item{sep}{A string indicating the separator to place between values in
observations. Defaults to " x "}

\item{col_sep}{A string indicating the separator to place between column
names. Defaults to \code{sep} with spaces replaced by underscores}
}
\value{
A modified \code{tbl} with an interaction column added
}
\description{
Description
}
\seealso{
\code{\link[=leaf_interactions]{leaf_interactions()}} to create multiple interactions at once
}
