% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaf_recode.R
\name{leaf_recode}
\alias{leaf_recode}
\title{Recode columns of a data frame}
\usage{
leaf_recode(tbl, code_tbl)
}
\arguments{
\item{tbl}{Data frame to modify}

\item{code_tbl}{A data frame of recoding values, either:
\itemize{
\item A data frame with three columns, \code{variable}, \code{code}, and \code{value}.
\code{variable} matches the column name to recode in \code{tbl}. \code{code} matches the
encoded values in \code{tbl}. \code{value} contains the values that each code
should be replaced with. (For example, see \code{leafpeepr::acs_codes_long}.)
\item A data frame with two adjacent columns for each column to be recoded in
\code{tbl}. One column contains encoded values and can have any name. The
other contains recoded values and has the same name as the column to
recode in \code{tbl}. (For example, see \code{leafpeepr::acs_codes}.)
}

Each code column can contain either literal codes or one-sided formulas.
(For example, compare \code{leafpeepr::acs_race_codes} and
\code{leafpeepr::acs_age_codes}.)}
}
\value{
A data frame with selected columns recoded
}
\description{
Recodes values in a data frame by joining with a data frame of recoded
values.
}
\examples{
# A data frame of encoded data
acs_md

# A data frame of recode values
acs_race_codes

#Recode
leaf_recode(acs_md, acs_race_codes)

# You can also specify recoding using formulas
acs_age_codes
leaf_recode(acs_md, acs_age_codes)

# You can use also use a data frame with recode values for multiple columns
# Either a wide data frame
acs_codes
leaf_recode(acs_md, acs_codes)

# Or a long data frame
acs_codes_long
leaf_recode(acs_md, acs_codes_long)
}
\seealso{
\code{\link[dplyr:case_when]{dplyr::case_when()}} to recode data using formulas
}
